refactor: 콘솔 게임 실행 흐름을 Application으로 통합
